modelIELTS/requirements.txt
# IELTS Scoring API Dependencies

# Core ML
torch>=2.0.0
transformers>=4.40.0
datasets>=2.14.0
evaluate>=0.4.0
numpy>=1.24.0

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Speech-to-Text (Whisper)
openai-whisper>=20230918

# Testing
requests>=2.31.0
pytest>=7.0.0
httpx>=0.24.0
